Witches House and Spell Updates
Began decorating witches house. Included Bed which you can sleep in to pass time till 8:00, a cauldron which you can craft potions at, and a small garden with some herbs to pick. Also adjusted spells so that mana would not be used if the spell couldn't be cast for lack of reagents.